+++ src/VGAudio.Benchmark/UtilityBenchmarks/BitReaderBenchmarks.cs
@@ -4,7 +4,7 @@ using BenchmarkDotNet.Attributes;
 using VGAudio.Utilities;
 
 
-namespace VGAudio.Benchmark.HelperBenchmarks
+namespace VGAudio.Benchmark.UtilityBenchmarks
 {
     public class BitReaderBenchmarks
     {
