+++ src/VGAudio.Benchmark/UtilityBenchmarks/InterleaveBenchmarks.cs
@@ -2,7 +2,7 @@
 using BenchmarkDotNet.Attributes;
 using VGAudio.Utilities;
 
-namespace VGAudio.Benchmark.HelperBenchmarks
+namespace VGAudio.Benchmark.UtilityBenchmarks
 {
     public class InterleaveBenchmarks
     {
